FIX: update example
